requirements.txt
absl-py==0.7.0
astor==0.7.1
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
gast==0.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-videointelligence==1.6.1
googleapis-common-protos==1.5.6
grpcio==1.18.0
h5py==2.9.0
idna==2.8
imutils==0.5.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.16.0
opencv-python==4.0.0.21
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
